Use docker run/stop directly to have more flexibility
By running containers via services we could not pass the amis.json via
volumes because when the container is spinned up we still have not run
the checkout code step.

Just directly spin up the container via `docker run', wait for it to be
ready via curl and then unconditionally stop it at the end.

No functional changes intended.