Move Terraform Makefile to terraform directory
All the targets are related to Terraform so it is better that it stays
in the terraform directory.

This is also needed because there will be Rego code that probably
deserves its own Makefile too.

(There is a `jobs.<job_id>.steps[*].working-directory` but let's stick
with `cd` that is more agnostic.)

NFCI.